+++ domain/users.go
@@ -2,14 +2,14 @@ package domain
 
 type UserUsecase interface {
 	CreateUser(user *User) error
-	GetUserIDByUsername(username string) (int, error)
-	GetUserByID(id int) (*User, error)
+	GetUserIDByUsername(username string) (int64, error)
+	GetUserByID(id int64) (*User, error)
 	GetUserByEmail(user *User) (*User, error)
 }
 
 type UserRepo interface {
 	CreateUser(user *User) error
-	GetUserIDByUsername(username string) (int, error)
-	GetUserByID(id int) (*User, error)
+	GetUserIDByUsername(username string) (int64, error)
+	GetUserByID(id int64) (*User, error)
 	GetUserByEmail(user *User) (*User, error)
 }
